Demo/services/tool-registry/requirements.txt
# ============================================
# Tool Registry - Python Dependencies
# ============================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Prometheus metrics
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
